backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
sqlalchemy==2.0.36
openai>=1.0.0
python-dotenv==1.0.1
pypdf==4.3.1
aiofiles==24.1.0
pydantic==2.9.2
psycopg2-binary==2.9.10
openpyxl==3.1.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.0
langgraph>=0.2.0
langgraph-checkpoint-postgres>=0.1.0
unstructured>=0.10.0
chromadb>=0.4.0
python-docx>=1.1.0
reportlab>=4.0.0

